pyproject.toml
[project]
name = "pixelpylot"
version = "0.1.0"
description = "Photography post-production workflow automation CLI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "photoshop-python-api>=0.24.1",
    "pillow>=11.2.1",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pip>=25.2",
]

[project.scripts]
pxp = "pixelpylot.cli:app"

[project.urls]
Homepage = "https://github.com/davidli218/pixelpylot"
Issues = "https://github.com/davidli218/pixelpylot/issues"
